Removed policy CMP0003 from BFITKPIPE CMakeLists
The CMake Policy CMP0003 was intended as part of a cross-platform hack
for making bfitk more portable on different platforms.

It was apparently part of the CMakeLists.txt before the rest of the hack
went into effect.  Thus when the hack was reverted, the policy setting
remained.

It has thus been removed.