Rename "BF-ITK" to "SCIFIO ImageIO"
With the underlying SCIFIO architecture now permissively licensed, we
can merge this ImageIO plugin into ITK core without GPL-related issues.

But as such, it makes more sense to use the "SCIFIO" name rather than
"Bio-Formats" since the code is not life-sciences-specific. This plugin
enables *any* SCIFIO-based I/O extensions (including Bio-Formats, but
potentially others as well) to be used with ITK.

This change also removes the pom.xml, which was used only for updating
the license blurbs. Since we are going to change the license anyway to
bring it in line with the rest of ITK, this POM is no longer necessary.