Define Java config using environment variables
We look at JAVA_HOME to find the location of the Java executable. If no
JAVA_HOME is defined, we assume "java" resides in the regular path.

We look for the SCIFIO JARs in the folder given by SCIFIO_PATH. (We
still need to provide a sensible default location if it is not set.)

We also still need to provide an environment variable for specifying
additional Java flags, especially so that we can override the default
maximum heap of 256MB.