Fix itkSCIFIOImageInfoTest argument handling
This commit works around the fact that the CMakeLists.txt (as written)
passes all arguments as a single quoted argument. There is almost
certainly a better way to do it in CMake, but I don't know how, so we
split the argument in C++ land instead.

This change also assigns the SCIFIO ImageIO to the reader properly.