Refactored ImageIOTest parameters
Changed itkSCIFIOImageIOTest to behave more like a *nix command line
application, with parameters passable through - abbreviated or
--explicit arguments.

The following can now be set:
* series range (for multi-series datasets)
* number of dimensions
* pixel type
* number of streaming divisions

This system should be easily extensible without concerns of parameter
order.

In the future, the RGB and Vector tests should be removed in favor of
additional flags for this test.