Pin to scifio-itk-bridge 1.1.0
We now deploy scifio-ikt-bridge artifacts to the imagej Maven repo.

The 1.0.0 artifact was not being rebuilt, and was just kind of left over
in the scifio-itk-bridge jenkins job - which could have been blown away
at any time if the job was cleaning.

We now pin only to artifacts that have been released to Maven.