BUG: Wrong installation directory and permissions for jre
When installing jre, permissions were not kept as in the downloaded
archive: executables did not have executable permissions after installation.
Additionally, the jre package that is downloaded was installed in
a subdirectory called 'jre' and also contains a directory called 'jre'
which created an installation directory with the following path:
.../jre/jre/...
When itkSCIFIOImageIO.cxx.in is configured, the configured installation
folder where to find 'java' is .../jre/... which lead to 'java'
not being found if JAVA_HOME was not specified manually to correct
this path issue.