BUG: Verifying that directory exists instead of java executable
To assess the directory in which the 'java' executable is installed,
SCIFIO was verifying that the given directory existed. However,
in certain cases, the directory existed, but the 'java' executable
did not exist. This could easily happened with the install directory
which by default is set to '/usr/local/' and that is likely to
exist. This would result in SCIFIO trying to use that path
to find the 'java' executable instead of searching in the PATH.
This bug is more easily visible on MacOS as SCIFIO assumes that
java should be installed on the system whereas for Windows and Linux
java is downloaded.